feat(checkout view): add checkout view for handling orders and payments
Implemented the checkout view in checkout/views.py. This view handles the process of creating an
order, linking it with the selected booking, and initializing a Stripe PaymentIntent. It includes
functionality for caching checkout data to enable a seamless payment experience for users.
Additionally, the view processes form data, validates the order, and redirects to a success page
upon successful completion